+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2-16] - 08-12-2022
+- Updates design for iOS 16 with new tables.
+- Further adds customizable toolbars.
+
 ## [1.2-15] - 14-09-2022 (no macOS release)
 - Removes "Tap to Copy" option and replaces it with a new toolbar item.
 - Adds support for customizable toolbars for iPad (iOS 16).